Tv shows (#3)
* Added tv shows full backend support

---------

+++ services/main-context/src/core/app/movieUpdateHandler.ts
@@ -71,6 +71,9 @@ export const handler = async (event: DynamoDBStreamEvent): Promise<void> => {
     for (const record of event.Records) {
       if (record.eventName === 'REMOVE') {
         let objectID = record.dynamodb?.Keys?.id.S;
+        if (objectID == undefined || ! /\S/.test(objectID)) {
+          throw new EmptyObjectIdError();
+        }
         await algoliaIndex.deleteBy({filters: `objectID: ${objectID}`});
         await cloudinary.api.delete_resources_by_prefix(objectID!);
         await emptyS3Directory(mediaAssetsS3Bucket, `${objectID}/`);
@@ -334,3 +337,5 @@ const tmdbPersonId2Person = {
 
 
 class FailedToGetMovieError extends Error {}
+
+class EmptyObjectIdError extends Error {}
